naim 0.11.7
This is a major release, with an accompanying bump in the version number
(from 0.11.6 to 0.11.7).  Much of the ECS and FireTalk backends have been
rewritten to clean up, optimize, and fix bugs in the code.  There is also
much more internal consistency-assurance code throughout naim, making it
easier for future bugs to be discovered and fixed.  I would encourage all
naim users to upgrade in order to make use of these new routines.