Backport naimng multisocket support into naim's firetalk.
This patch backports most of the naimng multisocket support into naim's
firetalk. It changes the API, and breaks some of the consumers (protocol
drivers). To keep the system functioning and compiling, I commented out all
protocol drivers but the one that I backported in that patch -- the IRC
driver. The IRC driver is renamed in that patch to TOC2 so that naim does
not crash on startup when it looks for a toc2 driver. nluapd is disabled
since the API has changed.