Fix proto_disconnected Lua hook.
Previously, there was an error in the setmetatable line for conn.groups, so the
Lua handler stopped there, and hence would not reset the session variables and
would not reset groups to be case insensitive. The end user impact should be a
fix for the assertion failures that you see after getting disconnected from the
server involuntarily. You will still get these assertion failures when you
/disconnect and re/connect.